vcachetuner: Improve annealing to be more correct
This change reframes the acceptance function around the correct
definition of temperature, where 0 means "don't accept worse outcomes",
and increasing temperature increases the possibility that a worse fit
wins.

The max temperature in this change is 0.1, which means that the
difference of 0.1 in fit (which is very large!) has a 50% chance of
winning. So the highest probability is probably very chaotic...